Better version of hand rules
